Moved speech_to_text_api.py in a module and removed import context from file with no main
